% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{compute_invgauss_ll_r}
\alias{compute_invgauss_ll_r}
\title{Compute inverse gaussian Log-Likelihood}
\usage{
compute_invgauss_ll_r(y, mu, gamma_val)
}
\arguments{
\item{y}{Response vector.}

\item{mu}{Mean (can be a vector or a matrix).}

\item{gamma_val}{Estimated dispersion}
}
\value{
The calculated log-likelihood.
}
\description{
Compute inverse gaussian Log-Likelihood
}
