% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{compute_ll}
\alias{compute_ll}
\title{Evaluate log-likelihoods}
\usage{
compute_ll(family, y, X, betas)
}
\arguments{
\item{family}{"gaussian", "binomial", "gamma", "inverse.gaussian", "poisson"}

\item{y}{Vector of y values. Required to have more than 1 observation}

\item{X}{Design matrix.}

\item{betas}{Either a vector of the beta values, or a matrix where each row corresponds to the beta vector to evaluate. Will be transposed so that it conforms to the dimensions of X}
}
\value{
A vector/scalar of Log-Likelihood value(s)
}
\description{
Helper function to evaluate log-likelihoods.
}
\examples{
\dontrun{
y <- rgamma(10, 3, 2)
X <- cbind(rep(1, 10), runif(10))
beta_eval_1 <- c(2.3, 12.4)
beta_eval_2 <- c(5, 18.9)
beta_eval_3 <- c(3.14, 22.2)
betas <- matrix(c(beta_eval_1, beta_eval_2, beta_eval_3), nrow = 3, byrow = TRUE)
compute_ll("gamma", y, X, betas)
}
}
