% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{compute_poisson_ll_r}
\alias{compute_poisson_ll_r}
\title{Compute Poisson Log-Likelihood}
\usage{
compute_poisson_ll_r(y, eta)
}
\arguments{
\item{y}{Response vector.}

\item{eta}{Linear predictor (can be a vector or a matrix).}
}
\value{
The calculated log-likelihood.
}
\description{
Compute Poisson Log-Likelihood
}
