% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{fit_glm_omp_r}
\alias{fit_glm_omp_r}
\title{Fit GLM using OpenMP via Rcpp}
\usage{
fit_glm_omp_r(X, y, mle_coefs, betas, family, num_threads, m, parallel, approx)
}
\arguments{
\item{X}{Predictor matrix.}

\item{y}{Response vector.}

\item{mle_coefs}{Maximum likelihood estimates of the coefficients.}

\item{betas}{Grid of beta values to evaluate.}

\item{family}{String indicating the error distribution.}

\item{num_threads}{The number of OpenMP threads to utilize.}

\item{m}{Number of evaluations per beta.}

\item{parallel}{Logical indicating whether to run in parallel.}

\item{approx}{Logical indicating whether to use elliptical approximation.}
}
\value{
Resulting output from the C++ GLM routine.
}
\description{
A wrapper to the underlying \code{fit_glm_omp_cpp} C++ implementation.
}
