% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immmmmmccccc.R
\name{generate_unit_matrix}
\alias{generate_unit_matrix}
\title{Generates all random numbers at once, so doesn't need to use the slow apply
Doesn't work if you pass in more than one column}
\usage{
generate_unit_matrix(n, d)
}
\description{
Generates all random numbers at once, so doesn't need to use the slow apply
Doesn't work if you pass in more than one column
}
