% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{get_CI}
\alias{get_CI}
\title{Returns a confidence interval}
\usage{
get_CI(glim_object, alpha = 0.05)
}
\arguments{
\item{glim_object}{Note that a 95\% confidence interval corresponds to alpha = .05}
}
\value{
A matrix of compatable beta values from the grid.
}
\description{
Note that a 95\% confidence interval corresponds to alpha = .05
}
