% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{glim}
\alias{glim}
\title{Generalized Linear Inferential Models (GLIM) Main Function}
\usage{
glim(
  X,
  y,
  family = "gaussian",
  m = 1000,
  approx = FALSE,
  appendix = FALSE,
  parallel = TRUE,
  intercept = TRUE,
  tol = 0.01,
  ...
)
}
\arguments{
\item{X}{Matrix of predictors.}

\item{y}{Vector of response variables.}

\item{family}{String denoting the exponential family. Choices are \code{"gaussian"}, \code{"binomial"}, \code{"gamma"}, \code{"poisson"}, \code{"inverse.gaussian"}.}

\item{m}{Number of samples/evaluations to perform (default \code{10000}).}

\item{approx}{Logical indicating whether to use the elliptical approximation (default \code{FALSE}).}

\item{parallel}{Logical indicating whether to process in parallel.}

\item{intercept}{Logical indicating whether to add an intercept term}

\item{betas}{A matrix (or column vector) of different beta values to evaluate the possibility over.}
}
\value{
A matrix of outputs or samples depending on whether the approximation is used.
}
\description{
Main wrapper function to fit a GLIM model.
}
