% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{glim_inner_prob_approx_samples}
\alias{glim_inner_prob_approx_samples}
\title{Generate Elliptical Approximation Samples (Inner Probability)}
\usage{
glim_inner_prob_approx_samples(
  X,
  y,
  family = "gaussian",
  mle_coefs,
  mle_val,
  m,
  parallel
)
}
\arguments{
\item{X}{Input predictor matrix.}

\item{y}{Response vector.}

\item{family}{A string indicating the error distribution. Default is \code{"gaussian"}.}

\item{mle_coefs}{Maximum likelihood estimates of the coefficients.}

\item{mle_val}{The log-likelihood value at the maximum likelihood estimates.}

\item{m}{Number of samples to generate.}

\item{parallel}{Logical indicating whether to use parallel processing.}
}
\value{
A matrix of generated samples based on the elliptical approximation.
}
\description{
Internal function called to generate samples when the elliptical approximation is used.
}
