% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{glim_raw}
\alias{glim_raw}
\title{Fits GLIM (Raw Implementation)}
\usage{
glim_raw(
  X,
  y,
  family = "gaussian",
  betas,
  mle_coefs,
  mle_val,
  m,
  parallel,
  approx
)
}
\arguments{
\item{X}{Input predictor matrix. If a data frame is provided, it will be coerced to a model matrix.}

\item{y}{Response vector.}

\item{family}{A string indicating the error distribution. Options include \code{"gaussian"}, \code{"poisson"}, \code{"gamma"}, \code{"binomial"}.}

\item{betas}{A grid of beta values to evaluate, where each column is a new beta vector. Not required if \code{approx = TRUE}.}

\item{mle_coefs}{Maximum likelihood estimates of the coefficients.}

\item{mle_val}{The log-likelihood value at the maximum likelihood estimates.}

\item{m}{Number of evaluations per beta (number of samples).}

\item{parallel}{Logical indicating whether to run in parallel (primarily for debugging).}

\item{approx}{Logical indicating whether to use the elliptical approximation.}
}
\value{
A matrix of evaluated possibility outputs.
}
\description{
Evaluates possibility for beta/dispersion values.
}
