% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{nobs.glim_object}
\alias{nobs.glim_object}
\title{nobs}
\usage{
\method{nobs}{glim_object}(object, ...)
}
\arguments{
\item{object}{A 'glim_object' from 'glim()'}

\item{...}{Additional arguments}
}
\value{
nobs
}
\description{
Returns the number of observations
}
