% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{plot.glim_object}
\alias{plot.glim_object}
\title{Plot}
\usage{
\method{plot}{glim_object}(x, ...)
}
\arguments{
\item{x}{A 'glim_object' from 'glim()'}

\item{...}{Currently unused}
}
\value{
Marginal plots for betas
}
\description{
Plotting for glim objects
}
