% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{prob2poss}
\alias{prob2poss}
\title{Probability to Possibility Transform}
\usage{
prob2poss(family, X, y, samples, the_compared_theta)
}
\arguments{
\item{family}{Either "guassian", "gamma", "inverse.gaussian", "poisson", "binomial"}

\item{X}{Design X matrix}

\item{y}{Vector of y values}

\item{samples}{Each column is a new beta vector sample}

\item{the_compared_theta}{Each row is vector of betas which we want the possibility}
}
\description{
From a set of samples (of beta values), return the possibility at any new beta value.
}
\examples{
\dontrun{
X <- cbind(intercept = 1, x1 = c(4, 3.3, 5))
y <- rgamma(3, shape = 2, scale = 1.4)

samples <- glim(y ~ X, family = "gamma", approx = TRUE)$samples

beta_1_grid <- seq(-2, 2, by = .2)
beta_2_grid <- seq(-1, 2, by = .2)
evaluated_new_theta <- expand.grid(beta_1_grid, beta_2_grid)
prob2poss("gamma", X, y, samples, evaluated_new_theta)
}
}
