% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{prob2poss_poisson}
\alias{prob2poss_poisson}
\title{Probability to Possibility Mapping for Poisson Regression}
\usage{
prob2poss_poisson(X, y, samples, the_compared_theta)
}
\arguments{
\item{X}{Predictor matrix.}

\item{y}{Response vector.}

\item{samples}{Matrix of simulated sample coefficients.}

\item{the_compared_theta}{The theta values to compare against.}
}
\value{
A vector of mapped possibility values.
}
\description{
Probability to Possibility Mapping for Poisson Regression
}
