% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{r_imvar}
\alias{r_imvar}
\title{Helper function for testing}
\usage{
r_imvar(
  X,
  y,
  xi,
  family,
  alpha,
  mle,
  mle_val,
  J_vectors,
  J_values,
  dispersion,
  tol,
  a_val,
  b_val,
  max_it,
  parallel,
  m
)
}
\arguments{
\item{X}{X}

\item{y}{y}

\item{xi}{xi}

\item{family}{family}

\item{alpha}{alpha}

\item{mle}{mle}

\item{mle_val}{mle_val}

\item{J_vectors}{J_vectors}

\item{J_values}{J_values}

\item{dispersion}{dispersion}

\item{tol}{tol}

\item{a_val}{a_val}

\item{b_val}{b_val}

\item{max_it}{max_it}

\item{parallel}{parallel}

\item{m}{m}
}
\value{
value
}
\description{
Helppp
}
