% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{vcov.glim_object}
\alias{vcov.glim_object}
\title{Returns the variance-covariance at the MLE}
\usage{
\method{vcov}{glim_object}(object, ...)
}
\arguments{
\item{object}{A 'glim_object' from 'glim()'}

\item{...}{Additional arguments}
}
\value{
A logLik class object
}
\description{
Extracts the variance-covariance matrix from the glim object.
}
