% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{coef.glim_object}
\alias{coef.glim_object}
\title{Extract the MLE coefficients}
\usage{
\method{coef}{glim_object}(object, ...)
}
\arguments{
\item{object}{A 'glim_object' from 'glim()'}

\item{...}{Additional arguments}
}
\value{
A vector of MLE coefficients
}
\description{
Extract the MLE coefficients
}
