% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{compute_gamma_ll_r}
\alias{compute_gamma_ll_r}
\title{Compute Gamma Log-Likelihood}
\usage{
compute_gamma_ll_r(y, eta, shape)
}
\arguments{
\item{y}{Response vector.}

\item{eta}{Linear predictor (can be a vector or a matrix).}

\item{shape}{The shape parameter for the gamma distribution.}
}
\value{
The calculated log-likelihood.
}
\description{
Compute Gamma Log-Likelihood
}
