% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{compute_gaussian_ll_r}
\alias{compute_gaussian_ll_r}
\title{Compute Gaussian Log-Likelihood}
\usage{
compute_gaussian_ll_r(y, mu, sigma)
}
\arguments{
\item{y}{Response vector.}

\item{mu}{Mean (can be a vector or a matrix).}

\item{sigma}{Estimated standard deviation}
}
\value{
The calculated log-likelihood.
}
\description{
Compute Gaussian Log-Likelihood
}
