% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{get_CI}
\alias{get_CI}
\title{Returns a confidence interval}
\usage{
get_CI(alpha, betas, possibilities)
}
\arguments{
\item{alpha}{Note that a 95\% confidence interval corresponds to alpha = .95}

\item{betas}{The grid of beta values which we are evaluating on}

\item{possibilities}{The vector of possibilities}
}
\value{
A matrix of compatable beta values from the grid.
}
\description{
Note that a 95\% confidence interval corresponds to alpha = .95
}
