% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{glim}
\alias{glim}
\title{Generalized Linear Inferential Models (GLIM) Main Function}
\usage{
glim(
  formula,
  data = NULL,
  family = "gaussian",
  m = 1000,
  approx = FALSE,
  appendix = FALSE,
  parallel = TRUE,
  tol = 0.01,
  ...
)
}
\arguments{
\item{formula}{Formula object to be interpreted}

\item{data}{Dataframe (if using)}

\item{family}{String denoting the exponential family. Choices are \code{"gaussian"}, \code{"binomial"}, \code{"gamma"}, \code{"poisson"}, \code{"inverse.gaussian"}.}

\item{m}{Number of samples/evaluations to perform (default \code{1000}).}

\item{approx}{Logical indicating whether to use the elliptical approximation (default \code{FALSE}).}

\item{appendix}{Bool indicating whether the appendix method is to be used.}

\item{parallel}{Logical indicating whether to process in parallel.}

\item{tol}{Double indicating the level of precision the approximation should get to}

\item{...}{Other arguments include a_val, b_val, max_it, betas, n_grid_evals, tol, and num_samps.}
}
\value{
If using an elliptical approximation, returns samples of parameter values. Else, returns a list containing the matrix of betas upon which it was evaluated, and a vector of the corresponding possibilities.
}
\description{
Main wrapper function to fit a GLIM model. Users can pass in their own grid of beta values to evaluate by setting betas = MATRIX.
A typical use would be MATRIX = as.matrix(expand.grid(beta_0_seq, beta_1_seq))
}
\details{
If using the elliptical approximation, can tweak the Robbins-Monroe algorithm by changing a_val, b_val, and max_it.
}
