% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{glim_raw}
\alias{glim_raw}
\title{Fits GLIM}
\usage{
glim_raw(
  X,
  y,
  family = "gaussian",
  betas,
  mle_coefs,
  mle_val,
  m,
  parallel,
  approx
)
}
\arguments{
\item{X}{input matrix}

\item{y}{response vector}

\item{family}{A string. Options include ("gaussian", "poisson", "gamma", "binomial")}

\item{betas}{If approximation is equal to false, this is not needed. Otherwise, this is a grid of beta values.
Each column is a new beta vector}

\item{mle_coefs}{Need to check if this is needed.}
}
\description{
Fits GLIM
}
