% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{print.glim_object}
\alias{print.glim_object}
\title{Print}
\usage{
\method{print}{glim_object}(output)
}
\arguments{
\item{object}{from 'glim()'}
}
\value{
IDK yet TODO
}
\description{
Printing for glim objects
}
