% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{print.glim_object}
\alias{print.glim_object}
\title{Print}
\usage{
\method{print}{glim_object}(x, ...)
}
\arguments{
\item{x}{A 'glim_object' from 'glim()'}

\item{...}{Supports differing alpha levels. EX: alpha = .05}
}
\value{
Print output
}
\description{
Printing for glim objects
}
