% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{prob2poss_logis}
\alias{prob2poss_logis}
\title{Probability to Possibility Mapping for Logistic Regression}
\usage{
prob2poss_logis(X, y, samples, the_compared_theta, intercept = TRUE)
}
\arguments{
\item{X}{Predictor matrix.}

\item{y}{Response vector.}

\item{samples}{Matrix of simulated sample coefficients.}

\item{the_compared_theta}{The theta values to compare against.}

\item{intercept}{Whether or not an intercept should be added to the design matrix}
}
\value{
A vector of mapped possibility values.
}
\description{
Probability to Possibility Mapping for Logistic Regression
}
