% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glim_logic.R
\name{scale_design_matrix}
\alias{scale_design_matrix}
\title{Scale the design matrix}
\usage{
scale_design_matrix(X)
}
\arguments{
\item{X}{design matrix}
}
\value{
Scaled design matrix (excludes intercepts and dummy variables)
}
\description{
Scales the design matrix for numerical stability when fitting
}
